executor: migrate test-infra to testify for testSuite1
